+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
-export const VERSION = "0.11.0";
+export const VERSION = "0.12.0";
 
 export type {
 	CreateComponentArgs,